Fixed: Set the port defined in sling.properties for the standalone server
When we change the port in sling.properties, let say for e.g. port is changed to 8084
And start the standalone server via command - percli server start

The server is started properly at 8084, but the message

'waiting for sling to start'

shown continuously on the console, and thus the Peregrine CMS has not launched on browser automatically.

*Issue*
The issue is checkIfRunning function is using hardcoded port 8080
and trying to access sling at 8080.

*Done following*

1. Implemented the getSlingPort function, which takes the path of sling.properties file and return the port defined in this file.

2. Call the getSlingPort function for the standalone server to get the port of sling, and set it via setPort method.

3. Moved the common code in getSlingPort function and call it wherever needed.